+++ database/migrations/2020_04_06_143718_create_inspection_tos_table.php
@@ -15,11 +15,11 @@ class CreateInspectionTosTable extends Migration
     {
         Schema::create('inspection_inspectors', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('inspection_id');
+            $table->unsignedBigInteger('inspection_id');
             $table->timestamps();
 
             $table->foreign('inspection_id')->references('id')->on('inspections')->onDelete('cascade')->onUpdate('cascade');
-            $table->unsignedInteger('inspector_id');
+            $table->unsignedBigInteger('inspector_id');
             $table->foreign('inspector_id')->references('id')->on('users')->onDelete('cascade')->onUpdate('cascade');
         });
     }
