+++ database/migrations/2020_04_15_085012_create_fcm_tokens_table.php
@@ -15,7 +15,7 @@ class CreateFcmTokensTable extends Migration
     {
         Schema::create('fcm_tokens', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger("user_id");
+            $table->unsignedBigInteger("user_id");
             $table->foreign("user_id")->references("id")->on("users")->onDelete("cascade");
             $table->text("token");
             $table->timestamps();
